Implemented CSV Simulation Exporter
Editor window capable of specific exporting of the simulation data.

One important thing to note is that depending on the users settings when opening the file in Excel, separating elements may need to be done with either semicolons or commas, and decimal places may need to be either periods or commas.